rename file getOptimalUpdateInterval
