+++ packages/react/ds-app-launchpad/src/ui/RelativeTime/utils/getOptimalUpdateInterval.ts
