+++ packages/react/ds-app-launchpad/src/ui/RelativeTime/utils/index.ts
@@ -1,3 +1,3 @@
 export { default as formatHumanTime } from "./formatHumanTime.js";
+export { default as getOptimalUpdateInterval } from "./getOptimalUpdateInterval.js";
 export { default as parseInstant } from "./parseInstant.js";
-export { default as getOptimalUpdateInterval } from "./optimalUpdateInterval.js";
